MKSBRYANS-166: fix

+++ mks_backend/services/documents/construction_document.py
@@ -23,7 +23,8 @@ class ConstructionDocumentService:
         construction_document = self.get_construction_document_by_id(id)
         file_storage_id = construction_document.idfilestorage
         self.repo.delete_construction_document(construction_document)
-        FilestorageRepository.delete_filestorage_by_id(file_storage_id)
+        if file_storage_id:
+            FilestorageRepository.delete_filestorage_by_id(file_storage_id)
 
     def update_construction_document(self, construction_document: ConstructionDocument) -> None:
         self.repo.update_construction_document(construction_document)
